refactor(filter): move FilterContents into Filter/components dir
